+++ layout/course.php
@@ -93,6 +93,9 @@ $regionmainsettingsmenu = $buildregionmainsettings ? $OUTPUT->region_main_settin
 $header = $PAGE->activityheader;
 $headercontent = $header->export_for_template($renderer);
 
+$theme = theme_config::load('stream');
+
+
 $templatecontext = [
     'sitename' => format_string($SITE->shortname, true, ['context' => context_course::instance(SITEID), "escape" => false]),
     'output' => $OUTPUT,
@@ -123,10 +126,18 @@ require_once(__DIR__ . '/includes/footer.php');
 // Include the content for scrollspy feature.
 require_once(__DIR__ . '/includes/scrollspy.php');
 
-$showcompletion = get_config('theme_stream', 'backtotopbutton');
-if ($showcompletion) {
+// Loads backtotop button. hribeiro dec2022.
+$backtotopbutton = $theme->settings->backtotopbutton;
+if ($backtotopbutton) {
     $PAGE->requires->js_call_amd('theme_stream/backtotop', 'init');
 }
 
+// Check for the option to print a course index heading.
+$courseindexheading = $theme->settings->courseindexheading;
+if ($courseindexheading != 0) {
+    $templatecontext['courseindexheading'] = format_string($this->page->course->$courseindexheading);
+} else {
+    $templatecontext['courseindexheading'] = null;
+}
 
 echo $OUTPUT->render_from_template('theme_stream/course', $templatecontext);